// src/app/shared/models/dataenty/request/PriceDataRequestForm.ts
export class PriceDataRequestForm {
  provinceId: number = null;
  monthTerm: number = null;
  yearTerm: number = null;
  productCode: string = null;
  productName: string = null;
  surveyName: string = null;
  sourceName: string = null;
  frequency: string = null;
  parentId: number = null;
  groupCode: string = null;
}